Added Seed Parameter to _sketch
Seed parameter will be of int format and will set the value for the internal Pseudo-random number generator, will have a default value of 0 with a special case ( -1 ) where the seed will be set to previous_seed + 1. The previous seed static variable is initialized with a value of 0 and gets updated to seed value passed by the user. When the user passes seed as -1 , the seed will be set to (previous_seed+1) and previous_seed will be incremented, hence giving us a changing seed for every figure produced.

Co-Authored-By: Oscar Gustafsson <8114497+oscargus@users.noreply.github.com>
Co-Authored-By: eudoxos <1029876+eudoxos@users.noreply.github.com>